fix(server): restore prod API — cold-start crash + checkout 503 (#1498)
Relocate Stripe mode classifier to @revealui/config (stops cold-start crash) + build optional Pro packages on Vercel so the function includes them (fixes checkout 503) + admin test skip null-safety. nft-verified; preview-verify checkout before promoting to main.
