+++ apps/server/package.json
@@ -64,7 +64,7 @@
     "test:coverage": "vitest run --coverage --coverage.reporter=json-summary --coverage.reporter=html --coverage.reporter=text",
     "test:watch": "vitest",
     "typecheck": "tsc --noEmit",
-    "vercel-build": "cd ../.. && pnpm turbo build --filter=server"
+    "vercel-build": "node scripts/vercel-build.mjs"
   },
   "type": "module"
 }
